\name{LauraeCE-package}
\alias{LauraeCE-package}
\docType{package}
\title{
Cross-Entropy R package for optimization
}
\description{The CEoptim package provides an optimization solver based on
  the Cross-Entropy method.
	The main function \code{\link{CEoptim}} can be used to solve
	multi-extremal optimization problems involving discrete, continuous,
	and mixed variables. In addition,    
	\code{CEoptim} implements linear constraints for continuous optimization.  
}
\details{
\tabular{ll}{
Package: \tab LauraeCE\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2017-08-10\cr
License: \tab GPL (>=2.0)\cr
LazyLoad: \tab yes\cr
}
}
\author{Laurae, Tim Benham, Qibin Duan, Dirk P. Kroese, Benoit Liquet <b.liquet@uq.edu.au>}
\references{
  Benham T.,  Duan Q., Kroese D.P., Liquet B. (2017) \bold{CEoptim}:
Cross-Entropy \R package for optimization. \emph{Journal of Statistical
  Software}, 76(8), 1-29.
}
\keyword{ package }
\seealso{
\code{\link{CEoptim}}
}

