\name{rtmvnorm}
\alias{rtmvnorm}
\title{Internal function used for generating truncated multivariate normal data}
\usage{
  rtmvnorm (N, mu, sigma, A, b,..., rhoThr=NULL, maxSample=NULL) 
}
\description{
  Internal function used for generating truncated multivariate normal data
}

\keyword{internal}
